move generated market files around
