+++ exchanges/binance-markets.json
@@ -1,5 +1,4 @@
-// File generated by update-binance.js, do not modify by hand!
-module.exports = {
+{
   "assets": [
     "NULS",
     "VEN",
