+++ exchanges/bitfinex-markets.json
@@ -1,5 +1,4 @@
-// File generated by update-bitfinex.js, do not modify by hand!
-module.exports = {
+{
   "assets": [
     "BTC",
     "LTC",
