+++ exchanges/kraken-markets.json
@@ -1,5 +1,4 @@
-// File generated by update-kraken.js, do not modify by hand!
-module.exports = {
+{
   "assets": [
     "BCH",
     "DASH",
