+++ exchanges/kraken.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const util = require('../core/util');
 const Errors = require('../core/error');
 const log = require('../core/log');
-const marketData = require('./data/krakenMarkets');
+const marketData = require('./kraken-markets.json');
 
 var Trader = function(config) {
   _.bindAll(this);
