+++ util/genMarketFiles/update-binance.js
