+++ util/genMarketFiles/update-bitfinex.js
