+++ util/genMarketFiles/update-kraken.js
